refactor(render): prefetch render pipeline buffer data
